+++ lib/chromic_pdf/pdf/browser/execution_error.ex
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: Apache-2.0
+
+defmodule ChromicPDF.Browser.ExecutionError do
+  @moduledoc """
+  Exception in interaction with the session pool.
+  """
+
+  defexception [:message]
+end